refactor: exportado componente de adminlist para melhor organização código
