+++ src/components/Pages/PageAdminSettings/PAdminSettings.vue
@@ -5,4 +5,3 @@
     </div>
   </div>
 </template>
-<script setup></script>
